v20141212_00
Speed Control Added
Change cputime to Uint32

+++ thedc1025/calculate/mhc_SCIA.c
@@ -1,7 +1,7 @@
 #include "mhc.h"
 #include "GlobalVariables.h"
 
-volatile Uint16 lastReciveTime = 0;
+volatile Uint32 lastReciveTime = 0;
 
 __interrupt void sciaTxFifoIsr(void)
 {
